requirements.txt
astroid==2.4.2
autopep8==1.5.3
click==7.1.2
flake8==3.8.3
Flask==1.1.2
flask-heroku==0.1.9
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
parso==0.7.0
passlib==1.7.2
pluggy==0.13.1
protobuf==3.12.1
psycopg2==2.8.5
pycodestyle==2.6.0
pydocstyle==5.0.2
pyflakes==2.2.0
pylint==2.5.3
python-dotenv==0.14.0
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
rope==0.17.0
six==1.14.0
snowballstemmer==2.0.0
SQLAlchemy==1.3.18
toml==0.10.1
ujson==1.35
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
yapf==0.30.0
